backend/requirements.txt
fastapi==0.115.5
uvicorn==0.27.1
sqlalchemy==2.0.25
asyncpg==0.29.0
python-dotenv==1.0.1
pydantic==2.10.3
aiogram==3.24.0
python-dotenv==1.0.1
aiohttp==3.10.5
PyJWT==2.9.0
